Fix problems with API functions causing crashes.
Before this, API functions returning lists would crash,
since it isn't possible to cast a `List<dynamic>` to a `List<String>`
via `dynamicList as List<String>` for example. The fix was to simply
move the casting out of `Session().call` and into each API function
individually, using `.cast<T>()` in functions returning `List`s and
the usual `as T` for the other functions. Thanks to
@bpilot on the [Dart SDK Gitter](https://gitter.im/dart-lang/sdk)
who came up with this fix and who helped a **lot**.

Also, regen bindings.